chore(charm): add test_nginx.py
moved NGINX ingress integration tests to own file
